+++ package.json
@@ -10,7 +10,6 @@
     "react-dom": "^16.6.0",
     "react-junctions": "^0.5.0",
     "react-scripts": "2.1.1",
-    "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1"
   },
   "homepage": "./",
