--- src/pages/Index/Index.js
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-
-class Index extends Component {
-    state = {  }
-    render() { 
-        return ( 
-            <div>
-                
-            </div> 
-        );
-    }
-}
- 
-export default Index;
